Assign shell-derived Makefile vars immediately
Makefile variables GITVER and BUILD create a shell subprocess to derive
their values.  By using the "=" assignment operator in the Makfile, this
shell subprocess was invoked each time the variable is used.  This could
(and did) lead to different files being compiled with different values
for these variables (especially BUILD).

To avoid this unwanted behavior, these variables are now assigned using
the ":=" immediate assignment operator so the shell subprocess is
invoked only once when the Makefile is parsed.